% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_script_addin.R
\name{create_new_script_addin}
\alias{create_new_script_addin}
\title{Add-in interactif pour creer un script standardise}
\usage{
create_new_script_addin()
}
\value{
Cree un fichier script dans le repertoire courant
}
\description{
Add-in interactif pour creer un script standardise
}
