% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-operators.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not In Operator}
\usage{
x \%notin\% table
}
\arguments{
\item{x, table}{objets à comparer, comme pour \code{\%in\%}.}
}
\value{
Un vecteur logique.
}
\description{
This operator is the negation of \code{\%in\%}.
}
\examples{
1 \%notin\% c(2, 3) # TRUE
1 \%notin\% c(1, 2) # FALSE
}
