% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Charge des données depuis une base via une requête SQL}
\usage{
loadData(query = character())
}
\arguments{
\item{query}{Chaîne de caractères contenant une requête SQL.}
}
\value{
Un objet \code{tibble} contenant les résultats de la requête.
}
\description{
Cette fonction établit une connexion à la base de données via ODBC (DSN = "dcshdb"), exécute la requête SQL passée en argument, récupère les résultats sous forme de tibble, puis ferme la connexion.
}
\examples{
\dontrun{
  req <- "SELECT * FROM ma_table"
  data <- loadData(req)
}

}
