% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightsCount.R
\name{nightsCount}
\alias{nightsCount}
\title{Hospital nights count}
\usage{
nightsCount(startDate = character(0), endDate = character(0))
}
\arguments{
\item{startDate}{A character string of length 12 representing the start date - time (e.g., "202305011200").}

\item{endDate}{A character string of length 12 representing the end date - time (e.g., "202305052315").}
}
\value{
An integer representing the number of nights between \code{startDate} and \code{endDate}.
If \code{startDate == endDate}, the result is 0. If \code{endDate < startDate}, the result is negative.
}
\description{
Computes the number of nights between two dates, based on the difference in days.
Dates must be provided as character strings in the same specified format.
}
\examples{
nightsCount("202305011150", "202305051155")
nightsCount("202305011230", "202305010152")
nightsCount("202305051300", "202305010000")
}
