% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptTemplate.R
\name{scriptTemplate}
\alias{scriptTemplate}
\title{Crée un script R structuré avec en-tête et sections prédéfinies}
\usage{
scriptTemplate(
  auteur = "Paul Majérus",
  titre = "",
  fileName = character(),
  version = "1",
  path = "."
)
}
\arguments{
\item{auteur}{Character. Nom de l’auteur à insérer dans l’en-tête. Par défaut \code{"Paul Majérus"}.}

\item{titre}{Character. Titre affiché en haut du fichier. Par défaut \code{""}.}

\item{fileName}{Character. Élément textuel à inclure dans le nom de fichier, nettoyé avec \code{make_clean_names(..., "small_camel")}. Peut être vide.}

\item{version}{Character. Numéro de version du script. Par défaut \code{"1"}.}

\item{path}{Character. Chemin du répertoire où le script sera enregistré. Par défaut \code{"."}.}
}
\value{
Aucun objet retourné, mais un fichier \code{.R} est créé sur disque. Affiche un message avec le chemin absolu du fichier.
}
\description{
Cette fonction génère automatiquement un fichier \code{.R} avec une structure standardisée,
incluant un en-tête (titre, auteur, date, version) et des sections de travail classiques.
Le fichier est nommé à partir de la date et du nom fourni, nettoyé au format camelCase.
}
\examples{
\dontrun{
create_new_script(auteur = "Paul Majérus",
                  titre = "Analyse des séjours hospitaliers",
                  fileName = "sejours_2025",
                  version = "2",
                  path = "R/scripts")
}

}
