% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_script_addin.R
\name{create_new_script_addin}
\alias{create_new_script_addin}
\title{Add-in interactif pour créer un script standardisé}
\usage{
create_new_script_addin()
}
\value{
Crée un fichier script dans le répertoire courant
}
\description{
Add-in interactif pour créer un script standardisé
}
